try a different casting

+++ jax/numpy/linalg.py
@@ -142,8 +142,8 @@ def pinv(a, rcond=None):
   a = np.conj(a)
   if rcond is None:
       max_rows_cols = max(a.shape[-2:])
-      rcond = 10. * max_rows_cols * np.finfo(a.dtype).eps
-  rcond = np.asarray(rcond, dtype=a.dtype)
+      rcond = np.float32(10) * max_rows_cols * np.finfo(a.dtype).eps
+  rcond = np.asarray(rcond)
   u, s, v = svd(a, full_matrices=False)
   # Singular values less than or equal to ``rcond * largest_singular_value`` 
   # are set to zero.
